dom/html/test/test_bug595429.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=595429
-->
<head>
  <title>Test for Bug 595429</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=595429">Mozilla Bug 595429</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 595429 **/

var fieldset = document.createElement("fieldset");

ok("name" in fieldset, "<fieldset> should have a name IDL attribute");

var testData = [
  "",
  " ",
  "foo",
  "foo bar",
];

is(fieldset.getAttribute("name"), null,
   "By default, name content attribute should be null");
is(fieldset.name, "",
   "By default, name IDL attribute should be the empty string");

for (var data of testData) {
  fieldset.setAttribute("name", data);
  is(fieldset.getAttribute("name"), data,
     "name content attribute should be " + data);
  is(fieldset.name, data, "name IDL attribute should be " + data);

  fieldset.setAttribute("name", "");
  fieldset.name = data;
  is(fieldset.getAttribute("name"), data,
     "name content attribute should be " + data);
  is(fieldset.name, data, "name IDL attribute should be " + data);
}

fieldset.removeAttribute("name");
is(fieldset.getAttribute("name"), null,
   "name content attribute should be null");
is(fieldset.name, "", "name IDL attribute should be the empty string");

</script>
</pre>
</body>
</html>
